MESH-316 | Support unarchive

+++ repository/src/main/java/org/apache/atlas/repository/migration/SoftDeletionProductMigrationService.java
@@ -16,6 +16,7 @@ import java.util.Set;
 
 import static org.apache.atlas.model.instance.AtlasEntity.Status.ACTIVE;
 import static org.apache.atlas.model.instance.AtlasEntity.Status.DELETED;
+import static org.apache.atlas.repository.Constants.MODIFICATION_TIMESTAMP;
 import static org.apache.atlas.repository.graph.GraphHelper.getStatus;
 
 public class SoftDeletionProductMigrationService {
@@ -57,7 +58,9 @@ public class SoftDeletionProductMigrationService {
                                 count++;
                                 totalUpdatedCount++;
                             }
-                        } else {
+                        }
+
+                        if (DELETED.equals(vertexStatus)) {
                             LOG.info("Removing edges for Archived Product: {}", productGuid);
                             boolean isCommitRequired = deleteEdgeForArchivedProduct(productVertex);
                             if (isCommitRequired) {
@@ -120,7 +123,7 @@ public class SoftDeletionProductMigrationService {
     private boolean deleteEdgeForArchivedProduct(AtlasVertex productVertex) {
         boolean isCommitRequired = false;
         try {
-            Long updatedTime = productVertex.getProperty("__modificationTimestamp", Long.class);
+            Long updatedTime = productVertex.getProperty(MODIFICATION_TIMESTAMP, Long.class);
             Iterator<AtlasEdge> existingEdges = productVertex.getEdges(AtlasEdgeDirection.BOTH).iterator();
 
             if (existingEdges == null || !existingEdges.hasNext()) {
@@ -130,7 +133,7 @@ public class SoftDeletionProductMigrationService {
 
             while (existingEdges.hasNext()) {
                 AtlasEdge edge = existingEdges.next();
-                Long modifiedEdgeTimestamp = edge.getProperty("__modificationTimestamp", Long.class);
+                Long modifiedEdgeTimestamp = edge.getProperty(MODIFICATION_TIMESTAMP, Long.class);
 
                 if (!updatedTime.equals(modifiedEdgeTimestamp)) {
                     LOG.info("Removing edge with different timestamp: {}", edge);
